efilesystem: shellhelper: move _Mount from imgutils.py to shellhelper.py
The rbind-based pseudo filesystem mounting for chroot() needs the
_Mount helper, but imgutils.py cannot be imported from shellhelper.py
without creating a cyclic dependency. Move _Mount there first.
