+++ elbepack/shellhelper.py
@@ -152,6 +152,51 @@ def _target_path(directory):
     return FALLBACK_PATH
 
 
+class _Mount:
+    # This is not using contextlib.contextmanager as it will be pass to our
+    # RPCAPTCache which uses the pickle serialization.
+    # The generator by contextlib.contextmanager is not compatible with pickle.
+    def __init__(self, device, target, *, bind=False, type=None, options=None, log_output=True,
+                 force_writable=False):
+        self.log_output = log_output
+        self.target = target
+
+        cmd = ['mount']
+        if bind:
+            cmd.append('--bind')
+
+        if options is not None:
+            cmd.extend(['-o', options])
+
+        if force_writable:
+            cmd.append('--rw')
+
+        if type is not None:
+            cmd.extend(['-t', type])
+
+        if device is None:
+            device = 'none'
+
+        cmd.extend([device, target])
+
+        self.cmd = cmd
+
+    def _run_cmd(self, cmd, *args, **kwargs):
+        if self.log_output:
+            do(cmd, *args, **kwargs)
+        else:
+            subprocess.run(cmd, *args, **kwargs)
+
+    def __enter__(self):
+        self._run_cmd(self.cmd)
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        self._run_cmd(['umount', self.target], check=False)
+
+
+mount = _Mount
+
+
 def chroot(directory, cmd, /, *, env_add=None, **kwargs):
     """chroot() - Wrapper around do().
